refactor(person): rename caver => person in reducers and actions
