+++ packages/web-app/src/reducers/CreatePerson.js
@@ -1,8 +1,8 @@
 import {
-  POST_CAVER,
-  POST_CAVER_FAILURE,
-  POST_CAVER_SUCCESS
-} from '../actions/CreateCaver';
+  POST_PERSON,
+  POST_PERSON_FAILURE,
+  POST_PERSON_SUCCESS
+} from '../actions/Person/CreatePerson';
 
 const initialState = {
   error: null,
@@ -12,18 +12,18 @@ const initialState = {
 
 const createCaver = (state = initialState, action) => {
   switch (action.type) {
-    case POST_CAVER:
+    case POST_PERSON:
       return {
         ...initialState,
         isLoading: true
       };
-    case POST_CAVER_SUCCESS:
+    case POST_PERSON_SUCCESS:
       return {
         ...initialState,
         isLoading: false,
         caver: action.caver
       };
-    case POST_CAVER_FAILURE:
+    case POST_PERSON_FAILURE:
       return {
         ...initialState,
         error: action.error
